requirements.txt
Flask==1.1.1
gunicorn==20.0.4
PyJWT==1.7.1
redis==3.3.11